[d2d] refactor text in own module
